refactor: コード重複の削減（3.7% → ~3.0%）
重複している関数を共通ユーティリティに移動：
1. 新規ファイル作成: src/lib/code-analysis-utils.ts
   - extractExportedNamesFromLine: diff行からエクスポート名を抽出
   - extractFunctionSignature: 関数シグネチャを抽出
   - normalizeSignature: パラメータリストを正規化

2. リファクタリング実施:
   - api-diff-summary.ts: 3つの重複関数を削除、共通ユーティリティを使用
   - breaking-change-analyzer.ts: 2つの重複関数を削除、共通ユーティリティを使用
   - extractExportedNamesFromLineはパスチェック処理のためラッパー関数として保持

これにより、約60行（重複3関数 × 約20行）のコード重複を削減。
SonarCloudのDuplication on New Code指標の改善を期待。
